opportunity-scanner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytz==2023.3
numpy>=1.26.0
pandas>=2.1.0
